fix(types): resolve usize→u32 regressions and regen bindings (#777)
All compile errors introduced by the usize→u32 type migration are fixed
across every feature set. Bindings regenerated with alef 0.15.70 after
rebasing onto main.

Changes:
- Replace all remaining `images:` PageContent construction sites with
  `image_indices: Vec<u32>` (extractors: image, pdf, docx, pptx, epub,
  jupyter, hwp, hwpx)
- Propagate u32 throughout: OcrElement.page_number, ChunkMetadata.first_page/
  last_page, PageBoundary.page_number, ExtractedImage.image_index, OcrMetadata
  table_count/rows/cols, PageStructure.total_count, PageHierarchy.block_count
- ChunkItem.first_page/last_page in api/types.rs: Option<usize> → Option<u32>
- paddle_ocr table_rows/cols: Option<usize> → Option<u32>
- ChunkMetadata.image_indices: Vec<usize> → Vec<u32> in alef.toml
- CHANGELOG: Vec<usize> → Vec<u32>; add Removed entry for PageContent.images
- docs/migration/v5.0-image-indices.md: before/after migration guide
- yaml_section.rs: descriptive comment → TODO(#963) with tracking issue
- swift.yml: suppress shellcheck SC2012 on intentional ls -dt

+++ crates/kreuzberg/src/api/types.rs
@@ -369,10 +369,10 @@ pub struct ChunkItem {
     pub total_chunks: usize,
     /// First page number (optional, for PDF chunking)
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub first_page: Option<usize>,
+    pub first_page: Option<u32>,
     /// Last page number (optional, for PDF chunking)
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub last_page: Option<usize>,
+    pub last_page: Option<u32>,
 }
 
 /// Version response.
